backend/requirements.txt
pipecat-ai[google,livekit,silero]>=0.0.102
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
livekit>=0.11.0
livekit-api>=0.6.0
python-dotenv>=1.0.0
